+++ waspc/data/Generator/templates/server/src/jobs/core/Job.ts
@@ -1,5 +1,3 @@
-import type { SubmittedJob } from "./SubmittedJob";
-
 /**
  * This is a definition of a job (think draft or invocable computation), not the running instance itself.
  * This can be submitted one or more times to be executed by some job executor via the same instance.
@@ -22,11 +20,33 @@ export function createJob<Input, Output extends object, Entities>(jobName: strin
     jobFn: async () => {
       throw new Error("Subclasses must implement this method");
     },
-    delay(...args: any[]) {
+    delay() {
       throw new Error("Subclasses must implement this method");
     },
-    async submit(...args: any[]) {
+    async submit() {
       throw new Error("Subclasses must implement this method");
     },
   } satisfies Job<Input, Output, Entities>;
+}
+
+/**
+ * This is the result of submitting a Job to some executor.
+ * It can be used by callers to track things, or call executor-specific subclass functionality.
+ */
+export interface SubmittedJob<Input, Output extends object, Entities> {
+  // job - The Job that submitted work to an executor.
+  job: Job<Input, Output, Entities>;
+  // jobId - A UUID for a submitted job in that executor's ecosystem.
+  jobId: string;
+  jobName: string;
+  executorName: string | symbol;
+}
+
+export function createSubmittedJob<Input, Output extends object, Entities>(job: Job<Input, Output, Entities>, jobId: string) {
+  return {
+    job,
+    jobId,
+    jobName: job.jobName,
+    executorName: job.executorName,
+  } satisfies SubmittedJob<Input, Output, Entities>;
 }
